fix docs

+++ Core/include/Acts/Seeding/SeedFilter.hpp
@@ -21,7 +21,7 @@
 namespace Acts {
 struct SeedConfQuantitiesConfig {
   // number of maximum top SPs in seed confirmation
-  size_t nTopSeedConf = std::numeric_limits<float>::max();
+  size_t nTopSeedConf = std::numeric_limits<int>::max();
   // number of high quality seeds in seed confirmation
   int numQualitySeeds = 0;
   // number of seeds that did not pass the quality confirmation but were still
